# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IELTS Hub API
  version: "1.0.0"
  summary: Authoring, exams, competitions, speaking practice and B2B analytics.
  description: |
    Single tagged document. Tags map 1:1 onto the internal modules defined in
    `docs/adr/0001-architecture.md` §4, so the API surface and the module graph
    stay legible against each other.

    ## Conventions

    * **Identifiers.** Every path parameter is an opaque `xid` (UUIDv7). Internal
      bigint keys are never exposed — sequential ids would turn the content
      library into a scraping API.
    * **Errors.** RFC 9457 `application/problem+json` everywhere. Validation
      failures return **every** problem at once, never the first.
    * **Pagination.** Cursor-based (`cursor` + `limit`). Offsets are unstable
      under concurrent inserts, and the authoring library sorts by `updated_at`.
    * **Idempotency.** `Idempotency-Key` is REQUIRED on attempt autosave, order
      creation and import commit; accepted on any other POST. A replayed key
      returns the stored response.
    * **Optimistic concurrency.** Draft content edits require `If-Match` with the
      ETag from the last read. Published content is immutable and rejects writes.
    * **Time.** RFC 3339 UTC. Anything time-critical also carries `server_now`;
      the client renders countdowns from the delta and never from its own clock.
    * **Realtime.** One WebSocket gateway. **Exam answers do NOT go over it** —
...

  contact: { name: Backend }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /test-versions/{xid}/validate:
    parameters: [{ $ref: '#/components/parameters/Xid' }]
    post:
      tags: [authoring-tests]
      summary: Run the publish gate without publishing
      description: |
        Returns **all 19 checks' findings at once** — never the first failure. Each
        finding carries a machine `code`, a JSON-pointer-ish `path` into the
        composition so the UI can deep-link to the broken item, and a `fix_hint`.
        The result is persisted, so an author can close the tab and come back.

        Requires **edit** authority on the version, not merely read. Findings quote
        the material they are about — `KEY_EXCEEDS_WORD_LIMIT` names the accepted
        answer verbatim — so this response can carry answer keys, and every student
        at a centre can read that centre's `org_private` content.
      responses:
        '200':
          description: Report. `passed` is false when any finding has severity `error`.
          content: { application/json: { schema: { $ref: '#/components/schemas/ValidationReport' } } }
        '403': { $ref: '#/components/responses/Problem' }
...
  /test-versions/{xid}/publish:
    parameters: [{ $ref: '#/components/parameters/Xid' }]
    post:
      tags: [authoring-tests]
      summary: Publish — runs the gate, then freezes the version
      description: |
        On success the version becomes immutable, the student-facing snapshot is
        materialized, and an audit record is written (`content.published`, naming
        the approving reviewer and the checksum they approved over). On failure
        returns **422 with every finding**, so the author fixes one round rather
        than nineteen.

        Teachers cannot publish unless the centre has enabled
        `settings.teacher_can_publish`.

        Where the centre has enabled `settings.require_review`, the version must
        also carry an approved review **over the content as it now stands**. The
        gate's 19 checks run first: "you have not been approved" is useless
        feedback on a test that would have failed validation anyway.
      responses:
        '200':
          description: Published.
          content: { application/json: { schema: { $ref: '#/components/schemas/TestVersion' } } }
        '403': { $ref: '#/components/responses/Problem' }
        '409':
          description: |
            `already_published` — this version is already live; create a new one.
            `review_required` — the centre requires a review and this version has
            none approved. `review_stale` — it was approved, then edited; the
            approval was over content that no longer exists.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
        '422': { $ref: '#/components/responses/ValidationFindings' }
...
  /question-versions/{xid}/keys:
    parameters: [{ $ref: '#/components/parameters/Xid' }]
    get:
      tags: [authoring-assets]
      summary: List key versions, newest first
      description: The full history of what this item was ever marked against.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AnswerKeyVersion' } }
    post:
      tags: [authoring-assets, regrade]
      summary: Fix an answer key — creates a new key version
      description: |
        **The key-fix path.** The question version stays frozen; a new
        `answer_key_versions` row supersedes the old one. Works on published
        content, which is exactly the point: a bad key must be fixable without
        invalidating what students already sat.

        When the item has been sat, the response includes a `regrade_preview` with
        the impact, and a `regrade_job_xid` created in `dry_run` state. **Nothing
        is regraded until you POST to `/regrades/{xid}/apply`.** If any affected
        attempt belongs to a competition, `regrade_preview.competition_impact` is
        populated and a platform-admin decision is required — a finished
        leaderboard never re-ranks itself.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AnswerKeyCreate' } } }
      responses:
        '201':
          description: New key version created; regrade staged but not applied.
          content:
            application/json:
              schema:
                type: object
                required: [key_version]
                properties:
                  key_version: { $ref: '#/components/schemas/AnswerKeyVersion' }
                  regrade_job_xid: { type: string, format: uuid }
                  regrade_preview: { $ref: '#/components/schemas/RegradeImpact' }
        '422': { $ref: '#/components/responses/ValidationFindings' }
...
  /imports:
    post:
      tags: [import-export]
      summary: Start an import — always a dry run first
      description: |
        Accepts DOCX, CSV or JSON. Every adapter produces the same **canonical
        Import JSON**, which is the contract that gets validated, diffed and
        committed; the adapters are replaceable and a weak DOCX parse degrades to
        a partial result with an error report rather than corrupting a test.

        Returns 202. Poll `GET /imports/{xid}` for the report.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              # `attestation` is required, not optional: "any uploader must affirm
              # the material is original or licensed, and that attestation is
              # logged with the upload". A bulk import is the likeliest route by
              # which a published paper arrives, so it is the last place to make
              # the affirmation optional.
              required: [file, format, attestation]
              properties:
                file: { type: string, format: binary }
                format: { type: string, enum: [docx, csv, json] }
                target_test_xid:
                  type: string
                  format: uuid
                  description: Import as a new version of an existing test — the offline round trip.
                attestation:
                  type: string
                  description: |
                    An `AttestationInput` object, JSON-encoded. Multipart fields
                    carry scalars, so unlike the JSON upload paths the object
                    travels here as a string.
      responses:
        '202':
          description: Parsing.
          content: { application/json: { schema: { $ref: '#/components/schemas/ImportJob' } } }
  /imports/{xid}:
    parameters: [{ $ref: '#/components/parameters/Xid' }]
    get:
      tags: [import-export]
      summary: Import status, error report and diff
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/ImportJob' } } }
  /imports/{xid}/commit:
    parameters: [{ $ref: '#/components/parameters/Xid' }]
    post:
      tags: [import-export]
      summary: Commit the dry run
      description: |
        Applies exactly the stored canonical JSON the author reviewed — not a
        re-parse, which could differ. Lands as a **draft** version for anyone who
        cannot publish, so bulk import can never become a publish bypass.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200':
          description: Committed.
          content: { application/json: { schema: { $ref: '#/components/schemas/ImportJob' } } }
        '409':
          description: Job is not in `validated` state, or the report contains errors.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access token. Refresh separately via `/auth/refresh`.
    paymeBasic:
      type: http
      scheme: basic
      description: Payme merchant key. Applies only to the Payme JSON-RPC endpoint.
  parameters:
    Xid:
      name: xid
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Opaque public identifier. Internal integer keys are never exposed.
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 128 }
      description: Client-generated. A replay returns the stored response.
...
  responses:
    Problem:
      description: RFC 9457 problem document.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationFindings:
      description: |
        Validation failed. `findings` carries **every** problem, never just the
        first — an author fixes one round rather than nineteen.
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/Problem' }
              - type: object
                properties:
                  findings: { type: array, items: { $ref: '#/components/schemas/Finding' } }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: 'https://api.example.uz/problems/publish-gate-failed' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Stable machine-readable code for client branching. }
        request_id: { type: string }
      additionalProperties: true
...
    ValidationReport:
      type: object
      required: [passed, findings, error_count, warning_count, run_at]
      properties:
        passed: { type: boolean, description: False when any finding has severity `error`. }
        findings: { type: array, items: { $ref: '#/components/schemas/Finding' } }
        error_count: { type: integer }
        warning_count: { type: integer }
        run_at: { type: string, format: date-time }
        duration_ms: { type: integer }
...
    TestVersion:
      type: object
      required: [xid, version_no, status]
      properties:
        xid: { type: string, format: uuid }
        version_no: { type: integer }
        status: { type: string, enum: [draft, in_review, published, archived] }
        title: { type: string }
        total_questions: { type: integer }
        max_raw: { type: number }
        band_map_version_xid: { type: [string, 'null'], format: uuid }
        published_at: { type: [string, 'null'], format: date-time }
        last_validation: { $ref: '#/components/schemas/ValidationReport' }
...
    AnswerKeyVersion:
      type: object
      required: [xid, version_no, key, is_current]
      properties:
        xid: { type: string, format: uuid }
        version_no: { type: integer }
        key: { type: object, additionalProperties: true }
        tolerance: { $ref: '#/components/schemas/Tolerance' }
        is_current: { type: boolean }
        reason: { type: string, enum: [initial, key_fix, clarification, import] }
        note: { type: [string, 'null'] }
        created_by: { $ref: '#/components/schemas/User' }
        created_at: { type: string, format: date-time }
        superseded_at: { type: [string, 'null'], format: date-time }
    AnswerKeyCreate:
      type: object
      required: [key, reason]
      properties:
        key:
          type: object
          additionalProperties: true
          description: |
            Shaped by the type's `key_schema`. For text types each slot carries
            `accept` (a list of accepted alternatives), optional `case_sensitive`,
            and optional per-key `normalizers` / `word_limit` overrides.
          example:
            slots:
              s1:
                accept: [fourteen, '14']
                case_sensitive: false
        tolerance: { $ref: '#/components/schemas/Tolerance' }
        reason: { type: string, enum: [initial, key_fix, clarification, import] }
        note: { type: string, description: Why the key changed. Shown in the audit trail and in regrade notices. }
...
    ImportJob:
      type: object
      required: [xid, status, source_format]
      properties:
        xid: { type: string, format: uuid }
        status: { type: string, enum: [received, parsing, validated, failed, committing, committed, cancelled] }
        source_format: { type: string, enum: [docx, csv, json] }
        report: { $ref: '#/components/schemas/ImportReport' }
        parse_error: { type: [string, 'null'] }
        committed_test_version_xid: { type: [string, 'null'], format: uuid }
        created_at: { type: string, format: date-time }
...